requirements.txt
# ===============================
# Core AI Orchestration
# ===============================

langchain==0.3.25
langchain-google-genai==2.1.4
google-genai==0.8.0
langchain-community==0.3.23

# ===============================
# Vector Database
# ===============================

chromadb==1.5.5

# ===============================
# News Scraping
# ===============================

newspaper4k==0.9.3.1
requests==2.32.3

# ===============================
# NLP
# ===============================

spacy==3.8.4

# ===============================
# Machine Learning
# ===============================

scikit-learn==1.6.1
numpy==2.2.4

# ===============================
# Frontend
# ===============================

streamlit==1.44.1

# ===============================
# Visualization
# ===============================

plotly==6.0.1

# ===============================
# Environment Variables
# ===============================

python-dotenv==1.1.0

# ===============================
# Utilities
# ===============================

tqdm==4.67.1

# ===============================
# Optional but Recommended
# ===============================

pandas==2.2.2